update api.pyからデータマッピング関連項目を分離して独立 責務の分離を目指す
